requirements.txt
gunicorn
psycopg2
django-heroku
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
CodeConvert==3.0.2
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==2.4.0
django-models-ext==1.1.10
django-six==1.0.4
djangorestframework==3.12.4
drf-yasg==1.20.0
furl==2.1.2
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
orderedmultidict==1.0.1
packaging==20.9
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
screen==1.0.1
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
TimeConvert==1.5.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
xlwt==1.3.0
